Use an absolute path to the conda environment
Updates the expected format of the `conda_environment` config parameter to be a
path relative to the top-level Snakefile instead of a path relative to the main
workflow files in `workflow/snakemake_rules/`. This change fixes a bug where
user-defined workflow rules in profiles (e.g.,
`wa_profile/download_preprocessed.smk`) are nested in a different directory
structure than the main workflow rules and cannot find the conda environment
based on a relative path.

This commit explicitly defines the `conda_environment` path as relative to the
top-level directory and then builds an absolute path from that relative path
such that all rules will reference the same file regardless of their directory
structure.